add update function for servers
